use u32 for hashmap

+++ lib/resources/src/retrieve/name.rs
@@ -16,6 +16,6 @@ impl<'a> NameRetrieve<'a> {
     /// Get a name by its sequence id
     #[inline]
     pub fn by_sequence(&self, seq_id: u32) -> Option<&'a Name> {
-        self.storage.names.get(seq_id as u64)
+        self.storage.names.get(seq_id)
     }
 }
